requirements.txt
google-api-python-client
pandas
plotly
python-dotenv
requests
python-pptx
Pillow
